refactor(analyzer)!: Move `PackageManagerDependencyHandler` to the root
Get rid of the empty intermediate `manager` package and move the file
next to `PackageManagerDependencyResult`.
