+++ analyzer/src/main/kotlin/AnalyzerResultBuilder.kt
@@ -19,7 +19,6 @@
 
 package org.ossreviewtoolkit.analyzer
 
-import org.ossreviewtoolkit.analyzer.managers.utils.PackageManagerDependencyHandler
 import org.ossreviewtoolkit.model.AnalyzerResult
 import org.ossreviewtoolkit.model.DependencyGraph
 import org.ossreviewtoolkit.model.DependencyGraphNavigator
